Split diagram handlers into sub-app classes
