+++ mainappsrc/AutoML.py
@@ -387,7 +387,6 @@ import builtins
 try:  # pragma: no cover - support direct module import
     from .user_manager import UserManager
     from .project_manager import ProjectManager
-    from .diagram_controller import DiagramController
     from .sotif_manager import SOTIFManager
 except Exception:  # pragma: no cover
     import os, sys
@@ -396,12 +395,10 @@ except Exception:  # pragma: no cover
     sys.path.append(os.path.dirname(base))
     from user_manager import UserManager
     from project_manager import ProjectManager
-    from diagram_controller import DiagramController
     from sotif_manager import SOTIFManager
 from user_manager import UserManager
 from project_manager import ProjectManager
 from cyber_manager import CyberSecurityManager
-from diagram_controller import DiagramController
 from cta_manager import ControlTreeManager
 from config.automl_constants import (
     dynamic_recommendations,
@@ -465,6 +462,26 @@ from analysis.utils import (
 )
 from analysis.safety_management import SafetyManagementToolbox, ACTIVE_TOOLBOX
 from analysis.causal_bayesian_network import CausalBayesianNetwork, CausalBayesianNetworkDoc
+try:  # pragma: no cover - support direct module import
+    from .style_subapp import StyleSubApp
+    from .tree_subapp import TreeSubApp
+    from .diagram_export_subapp import DiagramExportSubApp
+    from .use_case_diagram_subapp import UseCaseDiagramSubApp
+    from .activity_diagram_subapp import ActivityDiagramSubApp
+    from .block_diagram_subapp import BlockDiagramSubApp
+    from .internal_block_diagram_subapp import InternalBlockDiagramSubApp
+    from .control_flow_diagram_subapp import ControlFlowDiagramSubApp
+    from .version import VERSION
+except Exception:  # pragma: no cover
+    from style_subapp import StyleSubApp
+    from tree_subapp import TreeSubApp
+    from diagram_export_subapp import DiagramExportSubApp
+    from use_case_diagram_subapp import UseCaseDiagramSubApp
+    from activity_diagram_subapp import ActivityDiagramSubApp
+    from block_diagram_subapp import BlockDiagramSubApp
+    from internal_block_diagram_subapp import InternalBlockDiagramSubApp
+    from control_flow_diagram_subapp import ControlFlowDiagramSubApp
+    from version import VERSION
 try:  # pragma: no cover
     from .models.fault_tree_node import FaultTreeNode
 except Exception:  # pragma: no cover
@@ -507,28 +524,6 @@ def format_requirement(req, include_id=True):
 
 
 from pathlib import Path
-
-
-def get_version() -> str:
-    """Read the tool version from the first line of README.md.
-
-    The README is located alongside this file so we resolve the path relative
-    to ``__file__``.  This avoids returning ``"Unknown"`` when the current
-    working directory is different (e.g. when launching from another folder or
-    from an installed package).
-    """
-    try:
-        readme = Path(__file__).resolve().parent / "README.md"
-        with readme.open("r", encoding="utf-8") as f:
-            first_line = f.readline().strip()
-            if first_line.lower().startswith("version:"):
-                return first_line.split(":", 1)[1].strip()
-    except Exception:
-        pass
-    return "Unknown"
-
-
-VERSION = get_version()
 
 # Contact information for splash screen
 AUTHOR = "Miguel Marina"
@@ -662,197 +657,11 @@ class AutoMLApp:
         self.rc_dragged = False
         self.diagram_font = tkFont.Font(family="Arial", size=int(8 * self.zoom))
         self.style = ttk.Style()
-        try:
-            self.style.theme_use("clam")
-        except tk.TclError:
-            pass
-        self.style.configure(
-            "Treeview",
-            font=("Arial", 10),
-            background="#ffffff",
-            fieldbackground="#ffffff",
-            foreground="black",
-            borderwidth=1,
-            relief="sunken",
-        )
-        self.style.configure(
-            "Treeview.Heading",
-            background="#b5bdc9",
-            foreground="black",
-            relief="raised",
-        )
-        self.style.map(
-            "Treeview.Heading",
-            background=[("active", "#4a6ea9"), ("!active", "#b5bdc9")],
-            foreground=[("active", "white"), ("!active", "black")],
-        )
-        # ------------------------------------------------------------------
-        # Global color theme inspired by Windows classic / Windows 7
-        # ------------------------------------------------------------------
-        # Overall workspace background
-        root.configure(background="#f0f0f0")
-        # General widget colours
-        self.style.configure("TFrame", background="#f0f0f0")
-        self.style.configure("TLabel", background="#f0f0f0", foreground="black")
-        self.style.configure(
-            "TEntry", fieldbackground="#ffffff", background="#ffffff", foreground="black"
-        )
-        self.style.configure(
-            "TCombobox",
-            fieldbackground="#ffffff",
-            background="#ffffff",
-            foreground="black",
-        )
-        self.style.configure(
-            "TMenubutton", background="#e7edf5", foreground="black"
-        )
-        self.style.configure(
-            "TScrollbar",
-            background="#c0d4eb",
-            troughcolor="#e2e6eb",
-            bordercolor="#888888",
-            arrowcolor="#555555",
-            lightcolor="#eaf2fb",
-            darkcolor="#5a6d84",
-            borderwidth=2,
-            relief="raised",
-        )
-        # Apply the scrollbar styling to both orientations
-        for orient in ("Horizontal.TScrollbar", "Vertical.TScrollbar"):
-            self.style.configure(orient,
-                                background="#c0d4eb",
-                                troughcolor="#e2e6eb",
-                                bordercolor="#888888",
-                                arrowcolor="#555555",
-                                lightcolor="#eaf2fb",
-                                darkcolor="#5a6d84",
-                                borderwidth=2,
-                                relief="raised")
-        # Toolbox/LabelFrame titles
-        self.style.configure(
-            "Toolbox.TLabelframe",
-            background="#fef9e7",
-            bordercolor="#888888",
-            lightcolor="#fffef7",
-            darkcolor="#bfae6a",
-            borderwidth=1,
-            relief="raised",
-        )
-        self.style.configure(
-            "Toolbox.TLabelframe.Label",
-            background="#fef9e7",
-            foreground="black",
-            font=("Segoe UI", 10, "bold"),
-            padding=(4, 0, 0, 0),
-            anchor="w",
-        )
-        # Notebook (ribbon-like) title bars with beveled edges
-        self.style.configure(
-            "TNotebook",
-            background="#c0d4eb",
-            lightcolor="#eaf2fb",
-            darkcolor="#5a6d84",
-            borderwidth=2,
-            relief="raised",
-        )
-        self.style.configure(
-            "TNotebook.Tab",
-            background="#b5bdc9",
-            foreground="#555555",
-            borderwidth=1,
-            relief="raised",
-        )
-        self.style.map(
-            "TNotebook.Tab",
-            background=[("selected", "#4a6ea9"), ("!selected", "#b5bdc9")],
-            foreground=[("selected", "white"), ("!selected", "#555555")],
-        )
-        # Closable notebook shares the same appearance
-        self.style.configure(
-            "ClosableNotebook",
-            background="#c0d4eb",
-            lightcolor="#eaf2fb",
-            darkcolor="#5a6d84",
-            borderwidth=2,
-            relief="raised",
-        )
-        self.style.configure(
-            "ClosableNotebook.Tab",
-            background="#b5bdc9",
-            foreground="#555555",
-            borderwidth=1,
-            relief="raised",
-        )
-        self.style.map(
-            "ClosableNotebook.Tab",
-            background=[("selected", "#4a6ea9"), ("!selected", "#b5bdc9")],
-            foreground=[("selected", "white"), ("!selected", "#555555")],
-        )
-        # Mac-like capsule buttons
-        def _build_pill(top: str, bottom: str) -> tk.PhotoImage:
-            img = tk.PhotoImage(width=40, height=20)
-            # ``PhotoImage.put`` only accepts RGB colors. Some Tk builds
-            # (notably older Windows releases) mis-handle 8‑digit hex
-            # values used for transparency and instead raise a
-            # ``TclError``.  To keep the routine portable we fill the
-            # image with a solid RGB color and, where supported, mark that
-            # color as transparent.
-            img.put("#000000", to=(0, 0, 40, 20))
-            try:
-                img.transparency_set(0, 0, 0)
-            except Exception:
-                pass
-            radius = 10
-            t_r = int(top[1:3], 16)
-            t_g = int(top[3:5], 16)
-            t_b = int(top[5:7], 16)
-            b_r = int(bottom[1:3], 16)
-            b_g = int(bottom[3:5], 16)
-            b_b = int(bottom[5:7], 16)
-            for y in range(20):
-                ratio = y / 19
-                r = int(t_r * (1 - ratio) + b_r * ratio)
-                g = int(t_g * (1 - ratio) + b_g * ratio)
-                b = int(t_b * (1 - ratio) + b_b * ratio)
-                color = f"#{r:02x}{g:02x}{b:02x}"
-                for x in range(40):
-                    if x < radius:
-                        if (x - radius) ** 2 + (y - radius) ** 2 <= radius ** 2:
-                            img.put(color, (x, y))
-                    elif x >= 40 - radius:
-                        if (x - (40 - radius - 1)) ** 2 + (y - radius) ** 2 <= radius ** 2:
-                            img.put(color, (x, y))
-                    else:
-                        img.put(color, (x, y))
-            return img
-
-        self._btn_imgs = {
-            "normal": _build_pill("#fdfdfd", "#d2d2d2"),
-            "active": _build_pill("#eaeaea", "#c8c8c8"),
-            "pressed": _build_pill("#d0d0d0", "#a5a5a5"),
-        }
-        self.style.element_create(
-            "RoundedButton",
-            "image",
-            self._btn_imgs["normal"],
-            ("active", self._btn_imgs["active"]),
-            ("pressed", self._btn_imgs["pressed"]),
-            border=10,
-            sticky="nsew",
-        )
-        self.style.map(
-            "TButton",
-            relief=[("pressed", "sunken"), ("!pressed", "raised")],
-        )
-        # Navigation buttons used to scroll document tabs
+        self.style_app = StyleSubApp(root, self.style)
+        self.style_app.apply()
+        self._btn_imgs = self.style_app.btn_images
         self._init_nav_button_style()
-        # Increase notebook tab font/size so titles are fully visible
-        self.style.configure(
-            "TNotebook.Tab", font=("Arial", 10), padding=(10, 5), width=20
-        )
-        self.style.configure(
-            "ClosableNotebook.Tab", font=("Arial", 10), padding=(10, 5), width=20
-        )
+        self.tree_app = TreeSubApp()
         # style-aware icons used across tree views
         style_mgr = StyleManager.get_instance()
 
@@ -969,7 +778,12 @@ class AutoMLApp:
         self.user_manager = UserManager(self)
         self.project_manager = ProjectManager(self)
         self.cyber_manager = CyberSecurityManager(self)
-        self.diagram_controller = DiagramController(self)
+        self.diagram_export_app = DiagramExportSubApp(self)
+        self.use_case_diagram_app = UseCaseDiagramSubApp(self)
+        self.activity_diagram_app = ActivityDiagramSubApp(self)
+        self.block_diagram_app = BlockDiagramSubApp(self)
+        self.internal_block_diagram_app = InternalBlockDiagramSubApp(self)
+        self.control_flow_diagram_app = ControlFlowDiagramSubApp(self)
         self.sotif_manager = SOTIFManager(self)
         self.fmeda_manager = FMEDAManager(self)
         self.fmea_service = FMEAService(self)
@@ -1094,11 +908,11 @@ class AutoMLApp:
         review_menu.add_command(label="Merge Review Comments", command=self.merge_review_comments)
         review_menu.add_command(label="Compare Versions", command=self.compare_versions)
         architecture_menu = tk.Menu(menubar, tearoff=0)
-        architecture_menu.add_command(label="Use Case Diagram", command=self.diagram_controller.open_use_case_diagram)
-        architecture_menu.add_command(label="Activity Diagram", command=self.diagram_controller.open_activity_diagram)
-        architecture_menu.add_command(label="Block Diagram", command=self.diagram_controller.open_block_diagram)
-        architecture_menu.add_command(label="Internal Block Diagram", command=self.diagram_controller.open_internal_block_diagram)
-        architecture_menu.add_command(label="Control Flow Diagram", command=self.diagram_controller.open_control_flow_diagram)
+        architecture_menu.add_command(label="Use Case Diagram", command=self.open_use_case_diagram)
+        architecture_menu.add_command(label="Activity Diagram", command=self.open_activity_diagram)
+        architecture_menu.add_command(label="Block Diagram", command=self.open_block_diagram)
+        architecture_menu.add_command(label="Internal Block Diagram", command=self.open_internal_block_diagram)
+        architecture_menu.add_command(label="Control Flow Diagram", command=self.open_control_flow_diagram)
         architecture_menu.add_separator()
         architecture_menu.add_command(
             label="AutoML Explorer",
@@ -8222,186 +8036,20 @@ class AutoMLApp:
                                font=self.diagram_font)
 
     def save_diagram_png(self):
-        self.diagram_controller.save_diagram_png()
+        self.diagram_export_app.save_diagram_png()
 
     def on_treeview_click(self, event):
-        sel = self.analysis_tree.selection()
-        if not sel:
-            return
-        try:
-            node_id = int(self.analysis_tree.item(sel[0], "tags")[0])
-        except (IndexError, ValueError):
-            return
-        node = self.find_node_by_id_all(node_id)
-        if node:
-            self.open_page_diagram(node)
+        self.tree_app.on_treeview_click(self, event)
 
     def on_analysis_tree_double_click(self, event):
-        item = (
-            self.analysis_tree.identify_row(event.y)
-            if event is not None
-            else self.analysis_tree.focus()
-        )
-        if not item:
-            return
-        self.analysis_tree.focus(item)
-        tags = self.analysis_tree.item(item, "tags")
-        kind = tags[0] if tags else None
-        ident = tags[1] if tags and len(tags) > 1 else None
-        if kind in {"fmea", "fmeda", "hazop", "hara", "stpa", "threat", "fi2tc", "tc2fi", "jrev", "gov"} and ident is not None:
-            idx = int(ident)
-            if kind == "fmea":
-                self.show_fmea_table(self.fmeas[idx])
-            elif kind == "fmeda":
-                self.show_fmea_table(self.fmedas[idx], fmeda=True)
-            elif kind == "hazop":
-                self.open_hazop_window()
-                if hasattr(self, "_hazop_window"):
-                    doc = self.hazop_docs[idx]
-                    self._hazop_window.doc_var.set(doc.name)
-                    self._hazop_window.select_doc()
-            elif kind == "hara":
-                self.open_risk_assessment_window()
-                if hasattr(self, "_risk_window"):
-                    doc = self.hara_docs[idx]
-                    self._risk_window.doc_var.set(doc.name)
-                    self._risk_window.select_doc()
-            elif kind == "stpa":
-                self.open_stpa_window()
-                if hasattr(self, "_stpa_window"):
-                    doc = self.stpa_docs[idx]
-                    self._stpa_window.doc_var.set(doc.name)
-                    self._stpa_window.select_doc()
-            elif kind == "threat":
-                self.open_threat_window()
-                if hasattr(self, "_threat_window"):
-                    doc = self.threat_docs[idx]
-                    self._threat_window.doc_var.set(doc.name)
-                    self._threat_window.select_doc()
-            elif kind == "fi2tc":
-                self.open_fi2tc_window()
-                if hasattr(self, "_fi2tc_window"):
-                    doc = self.fi2tc_docs[idx]
-                    self._fi2tc_window.doc_var.set(doc.name)
-                    self._fi2tc_window.select_doc()
-            elif kind == "tc2fi":
-                self.open_tc2fi_window()
-                if hasattr(self, "_tc2fi_window"):
-                    doc = self.tc2fi_docs[idx]
-                    self._tc2fi_window.doc_var.set(doc.name)
-                    self._tc2fi_window.select_doc()
-            elif kind == "jrev":
-                if 0 <= idx < len(getattr(self, "joint_reviews", [])):
-                    review = self.joint_reviews[idx]
-                    self.review_data = review
-                    self.open_review_document(review)
-                    self.open_review_toolbox()
-            elif kind == "gov":
-                self.open_management_window(idx)
-        elif kind == "gsn" and ident is not None:
-            diag = getattr(self, "gsn_diagram_map", {}).get(ident)
-            if diag:
-                self.open_gsn_diagram(diag)
-        elif kind == "gsnmod":
-            self.manage_gsn()
-        elif kind == "reqs":
-            self.show_requirements_editor()
-        elif kind == "reqexp":
-            self.show_requirements_explorer()
-        elif kind == "sg":
-            self.show_product_goals_editor()
-        elif kind == "fta" and ident is not None:
-            te = next((t for t in self.top_events if t.unique_id == int(ident)), None)
-            if te:
-                self.diagram_mode = "FTA"
-                self.ensure_fta_tab()
-                self.doc_nb.select(self.canvas_tab)
-                self.open_page_diagram(te)
-        elif kind == "cta" and ident is not None:
-            te = next((t for t in getattr(self, "cta_events", []) if t.unique_id == int(ident)), None)
-            if te:
-                self.diagram_mode = "CTA"
-                self.ensure_fta_tab()
-                self.doc_nb.select(self.canvas_tab)
-                self.open_page_diagram(te)
-        elif kind == "paa" and ident is not None:
-            te = next((t for t in getattr(self, "paa_events", []) if t.unique_id == int(ident)), None)
-            if te:
-                self.diagram_mode = "PAA"
-                self.ensure_fta_tab()
-                self.doc_nb.select(self.canvas_tab)
-                self.open_page_diagram(te)
-        elif kind == "safetycase":
-            self.manage_safety_cases()
-        elif kind == "safetyconcept":
-            self.show_safety_concept_editor()
-        elif kind == "itemdef":
-            self.show_item_definition_editor()
-        elif kind == "arch":
-            self.open_arch_window(ident)
-        elif kind == "pkg":
-            self.manage_architecture()
-        else:
-            parent = item
-            while parent:
-                if (
-                    self.analysis_tree.item(parent, "text")
-                    == "Safety & Security Governance Diagrams"
-                ):
-                    self.manage_safety_management()
-                    return
-                parent = self.analysis_tree.parent(parent)
+        self.tree_app.on_analysis_tree_double_click(self, event)
 
     def on_analysis_tree_right_click(self, event):
-        iid = self.analysis_tree.identify_row(event.y)
-        if not iid:
-            return
-        self.analysis_tree.selection_set(iid)
-        self.analysis_tree.focus(iid)
-        menu = tk.Menu(self.analysis_tree, tearoff=0)
-        menu.add_command(label="Rename", command=self.rename_selected_tree_item)
-        menu.tk_popup(event.x_root, event.y_root)
+        self.tree_app.on_analysis_tree_right_click(self, event)
 
     def on_analysis_tree_select(self, _event):
         """Update property view when a tree item is selected."""
-        if not hasattr(self, "prop_view"):
-            return
-        item = self.analysis_tree.focus()
-        if not item:
-            return
-        tags = self.analysis_tree.item(item, "tags")
-        name = self.analysis_tree.item(item, "text")
-        meta = {"Name": name}
-        if tags:
-            meta["Type"] = tags[0]
-            if len(tags) > 1:
-                ident = tags[1]
-                meta["ID"] = ident
-                repo = SysMLRepository.get_instance()
-                elem = repo.elements.get(ident)
-                if elem:
-                    meta.update(
-                        {
-                            "Type": elem.elem_type,
-                            "Author": getattr(elem, "author", ""),
-                            "Created": getattr(elem, "created", ""),
-                            "Modified": getattr(elem, "modified", ""),
-                            "ModifiedBy": getattr(elem, "modified_by", ""),
-                        }
-                    )
-                else:
-                    diag = repo.diagrams.get(ident)
-                    if diag:
-                        meta.update(
-                            {
-                                "Type": diag.diag_type,
-                                "Author": getattr(diag, "author", ""),
-                                "Created": getattr(diag, "created", ""),
-                                "Modified": getattr(diag, "modified", ""),
-                                "ModifiedBy": getattr(diag, "modified_by", ""),
-                            }
-                        )
-        self.show_properties(meta=meta)
+        self.tree_app.on_analysis_tree_select(self, _event)
 
     def show_properties(self, obj=None, meta=None):
         """Display metadata for *obj* or *meta* dictionary in the properties tab."""
@@ -8457,125 +8105,7 @@ class AutoMLApp:
                         self.status_meta_vars[k].set(v)
 
     def rename_selected_tree_item(self):
-        item = self.analysis_tree.focus()
-        tags = self.analysis_tree.item(item, "tags")
-        if len(tags) != 2:
-            return
-        kind, ident = tags[0], tags[1]
-        repo = SysMLRepository.get_instance()
-        current = ""
-        node = None
-        if kind in {"fmea", "fmeda", "hazop", "hara", "fi2tc", "tc2fi", "jrev"}:
-            idx = int(ident)
-            if kind == "fmea":
-                current = self.fmeas[idx]["name"]
-            elif kind == "fmeda":
-                current = self.fmedas[idx]["name"]
-            elif kind == "hazop":
-                current = self.hazop_docs[idx].name
-            elif kind == "hara":
-                current = self.hara_docs[idx].name
-            elif kind == "fi2tc":
-                current = self.fi2tc_docs[idx].name
-            elif kind == "tc2fi":
-                current = self.tc2fi_docs[idx].name
-            elif kind == "jrev":
-                current = self.joint_reviews[idx].name
-        elif kind == "gsn":
-            diag = getattr(self, "gsn_diagram_map", {}).get(ident)
-            if not diag:
-                return
-            current = diag.root.user_name
-        elif kind == "gsnmod":
-            module = getattr(self, "gsn_module_map", {}).get(ident)
-            if not module:
-                return
-            current = module.name
-        elif kind == "arch":
-            diag = repo.diagrams.get(ident)
-            current = diag.name if diag else ""
-        elif kind == "gov":
-            idx = int(ident)
-            current = self.management_diagrams[idx].name
-        elif kind == "fta":
-            node = next((t for t in self.top_events if t.unique_id == int(ident)), None)
-            current = node.user_name if node else ""
-        elif kind == "pkg":
-            pkg = repo.elements.get(ident)
-            current = pkg.name if pkg else ""
-        else:
-            return
-        new = simpledialog.askstring("Rename", "Enter new name:", initialvalue=current)
-        if not new:
-            return
-        if kind == "fmea":
-            old = self.fmeas[idx]["name"]
-            self.fmeas[idx]["name"] = new
-            self.safety_mgmt_toolbox.rename_document("FMEA", old, new)
-        elif kind == "fmeda":
-            doc = self.fmedas[idx]
-            self.fmeda_manager.rename_fmeda(doc, new)
-        elif kind == "hazop":
-            old = self.hazop_docs[idx].name
-            self.hazop_docs[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("HAZOP", old, new)
-        elif kind == "hara":
-            old = self.hara_docs[idx].name
-            self.hara_docs[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("Risk Assessment", old, new)
-        elif kind == "fi2tc":
-            old = self.fi2tc_docs[idx].name
-            self.fi2tc_docs[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("FI2TC", old, new)
-        elif kind == "tc2fi":
-            old = self.tc2fi_docs[idx].name
-            self.tc2fi_docs[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("TC2FI", old, new)
-        elif kind == "fta":
-            node = next((t for t in self.top_events if t.unique_id == int(ident)), None)
-            if node:
-                old = node.user_name
-                node.user_name = new
-                analysis = (
-                    "Prototype Assurance Analysis"
-                    if getattr(self, "diagram_mode", "") == "PAA"
-                    else "FTA"
-                )
-                self.safety_mgmt_toolbox.rename_document(analysis, old, new)
-        elif kind == "arch" and repo.diagrams.get(ident):
-            repo.diagrams[ident].name = new
-        elif kind == "gov":
-            self.management_diagrams[idx].name = new
-        elif kind == "gsn":
-            diag = self.gsn_diagram_map.get(ident)
-            if diag:
-                diag.root.user_name = new
-        elif kind == "gsnmod":
-            module = self.gsn_module_map.get(ident)
-            if module:
-                module.name = new
-        elif kind == "jrev":
-            if any(r.name == new for r in self.reviews if r is not self.joint_reviews[idx]):
-                messagebox.showerror("Review", "Name already exists")
-                return
-            old = self.joint_reviews[idx].name
-            self.joint_reviews[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("Joint Review", old, new)
-        elif kind == "fta" and node:
-            old = node.name
-            node.user_name = new
-            if hasattr(self, "safety_mgmt_toolbox"):
-                analysis = (
-                    "Prototype Assurance Analysis"
-                    if getattr(self, "diagram_mode", "") == "PAA"
-                    else "FTA"
-                )
-                self.safety_mgmt_toolbox.rename_document(analysis, old, node.name)
-        elif kind == "pkg" and repo.elements.get(ident):
-            repo.elements[ident].name = new
-        self.update_views()
-        if hasattr(self, "_arch_window") and self._arch_window.winfo_exists():
-            self._arch_window.populate()
+        self.tree_app.rename_selected_tree_item(self)
 
     def on_tool_list_double_click(self, event):
         lb = event.widget
@@ -17096,19 +16626,19 @@ class AutoMLApp:
         return create_icon(shape, color)
 
     def open_use_case_diagram(self):
-        self.diagram_controller.open_use_case_diagram()
+        self.use_case_diagram_app.open()
 
     def open_activity_diagram(self):
-        self.diagram_controller.open_activity_diagram()
+        self.activity_diagram_app.open()
 
     def open_block_diagram(self):
-        self.diagram_controller.open_block_diagram()
+        self.block_diagram_app.open()
 
     def open_internal_block_diagram(self):
-        self.diagram_controller.open_internal_block_diagram()
+        self.internal_block_diagram_app.open()
 
     def open_control_flow_diagram(self):
-        self.diagram_controller.open_control_flow_diagram()
+        self.control_flow_diagram_app.open()
 
     def manage_architecture(self):
         if hasattr(self, "_arch_tab") and self._arch_tab.winfo_exists():
